Attach degrees of freedom to electrodes
Each Fixed/FloatingPotentialElectrode now holds a degree of freedom associated with either electric potential or surface charge density.
The degree of freedom is a view (subarray), so that it can be read with an up-to-date value, and can be updated using .=
To make argument names consisted, I've changed V to ϕ as the prescribed value of electric potential.