Record ε0 in PoissonSolver, so we can scale right hand side vector
Both surface charge density and space charge density have to be scaled with ε0 when putting in RHS. To avoid having
to scale it by the user, the whole RHS vector could be scaled while solving for the field.
Only degrees of freedom with fixed electric potential shouldn't be scaled. To avoid having to remember the nodes,
allow only zero-value Dirichlet Boundary Condition (in the following commit).