Refactoring of SurfaceTracker
Split SurfaceTracker into more manageable pieces (build, check, hit, track, wrap).
Remove functions defining if Surface is absorbing or reflecting. Now it is controlled by
the surface on each hit! when it can call absorbed! or scattered! function. It is because
this way the surface can absorb some particles and reflect the rest (maybe on random?).