Change diagnostics module to depend on OpenPMD standard
This way, the result data are easier to share, better structured and allow interoperability.
It is possible now to analyze the data in Jupyter Notebook using openPMD-viewer for Python.
Later, the files could be regarded as checkpoints for restarting simulations.

Reworked Chemistry and ParticleInCell modules to match the new API of Diagnostics

There is still missing integration with Circuit as 0D time series has to be figured out. According to the standard, they should be stored *outside* of /data/iteration.

Adjust 01_single_electron problem to new API