Keep track of DOFs for source term
When a Dirichlet Boundary Condition is applied, the right hand side vector hold only the value on the boundary, so the source term SHOULD NOT be added to it. Now, there are two degree of freedom sets: ϕ and ρ. ϕ can be used to extract the full electric potential field, while ρ keeps track of DOFs where the source term should be applied (so all in ϕ except the ones where the DBC has been applied). ϕ is stored as an array while ρ is a vector so it can be used to select entries from source term vector f and RHS vector b.