Modify gather/scatter functions to work in cylindrical coordinate system
There is a bug here. In particle_to_grid the u(i,j) is weighted with 1/ΔV, because it is being used for number density (particle_to_grid is fed with number of real particles in macroparticles, and if we divide each by ΔV we get the number density). It won't work for any property other than number density.

If it is resolved, particle_to_grid and grid_to_particle will look the same for all UniformGrids. Also, volume calculation for each particle is time consuming, so some smart precomputed array is needed.